requirements.txt
blinker==1.6.2
cachelib==0.10.2
click==8.1.4
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
Flask-Session==0.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector-python==8.0.33
protobuf==3.20.3
Werkzeug==2.3.6
